fix(): organizing deps, adding javadocs
