+++ function/function-samples/sample-function-book-aws/src/main/java/io/americanexpress/function/book/function/BookFunction.java
@@ -35,49 +35,90 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
+/**
+ * {@code BookFunction} contains functions for crud operations /reactive/book.
+ */
 @Configuration
 public class BookFunction {
 
     private final String BOOK_NOT_FOUND = "Book Not Found";
 
+    /**
+     * Used for crud operations on book table in cassandra database.
+     */
     private final BookRepository bookRepository;
 
+    /**
+     * Instantiates a new Book function.
+     *
+     * @param bookRepository the book repository
+     */
     public BookFunction(BookRepository bookRepository) {
         this.bookRepository = bookRepository;
     }
 
+    /**
+     * Create function.
+     *
+     * @return the function
+     */
     @Bean
-    public Function<Message<CreateBookRequest>, Mono<CreateBookResponse>> create() {
+    public Function<Message<CreateBookRequest>, Mono<CreateBookResponse>> create() { //406
         return request -> bookRepository.save(BookEntityCreator.create(request.getPayload().getTitle(), request.getPayload().getAuthor(), 1))
                     .map(bookEntity -> new CreateBookResponse());
     }
 
+    /**
+     * Create book consumer.
+     *
+     * @return the consumer
+     */
     @Bean
-    public Consumer<Message<CreateBookRequest>> createBook() {
+    public Consumer<Message<CreateBookRequest>> createBook() { //202
         return request -> bookRepository.save(BookEntityCreator.create(request.getPayload().getTitle(), request.getPayload().getAuthor(), 1));
     }
 
+    /**
+     * Read function.
+     *
+     * @return the function
+     */
     @Bean
-    public Function<Message<ReadBookRequest>, Mono<ReadBookResponse>> read() {
+    public Function<Message<ReadBookRequest>, Mono<ReadBookResponse>> read() { //200
         return request -> bookRepository.findByTitleAndAuthor(request.getPayload().getTitle(), request.getPayload().getAuthor())
                 .map(ReadBookResponseCreator::create);
     }
 
+    /**
+     * Get supplier.
+     *
+     * @return the supplier
+     */
     @Bean
-    public Supplier<Flux<ReadBookResponse>> get() {
+    public Supplier<Flux<ReadBookResponse>> get() { //200
         return () -> bookRepository.findAll().map(ReadBookResponseCreator::create);
     }
 
+    /**
+     * Update consumer.
+     *
+     * @return the consumer
+     */
     @Bean
-    public Consumer<Message<UpdateBookRequest>> update() {
+    public Consumer<Message<UpdateBookRequest>> update() { //202
         return request -> bookRepository.findByTitleAndAuthor(request.getPayload().getTitle(), request.getPayload().getAuthor())
                 .switchIfEmpty(Mono.error(new ResponseStatusException(HttpStatus.NOT_FOUND, BOOK_NOT_FOUND)))
                 .map(book -> updateBook(book, request.getPayload().getNumberOfCopies()))
                 .flatMap(bookRepository::save);
     }
 
+    /**
+     * Update book function.
+     *
+     * @return the function
+     */
     @Bean
-    public Function<Message<UpdateBookRequest>, Mono<CreateBookResponse>> updateBook() {
+    public Function<Message<UpdateBookRequest>, Mono<CreateBookResponse>> updateBook() { //406
         return request -> bookRepository.findByTitleAndAuthor(request.getPayload().getTitle(), request.getPayload().getAuthor())
                 .switchIfEmpty(Mono.error(new ResponseStatusException(HttpStatus.NOT_FOUND, BOOK_NOT_FOUND)))
                 .map(book -> updateBook(book, request.getPayload().getNumberOfCopies()))
@@ -85,13 +126,25 @@ public class BookFunction {
                 .map(bookEntity -> new CreateBookResponse());
     }
 
+    /**
+     * Update book entity given with number of copies.
+     *
+     * @param book the bookEntity to be updated
+     * @param numOfCopies the number of copies
+     * @return the bookEntity created
+     */
     private BookEntity updateBook(BookEntity book, int numOfCopies){
         if (book != null) book.setNumberOfCopies(numOfCopies);
         return book;
     }
 
+    /**
+     * Delete consumer.
+     *
+     * @return the consumer
+     */
     @Bean
-    public Consumer<Message<BookRequest>> delete() {
+    public Consumer<Message<BookRequest>> delete() { //202
         return request -> bookRepository.findByTitleAndAuthor(request.getPayload().getTitle(), request.getPayload().getAuthor())
                 .switchIfEmpty(Mono.error(new ResponseStatusException(HttpStatus.NOT_FOUND, BOOK_NOT_FOUND)))
                 .flatMap(bookRepository::delete);
