+++ function/function-samples/sample-function-book-aws/src/main/java/io/americanexpress/function/book/function/helper/ReadBookResponseCreator.java
@@ -21,6 +21,9 @@ import io.americanexpress.function.book.model.ReadBookResponse;
  */
 public class ReadBookResponseCreator {
 
+    /**
+     * Private constructor to prevent instantiation of BookEntityCreator.
+     */
     private ReadBookResponseCreator() {}
 
     /**
