+++ pom.xml
@@ -47,6 +47,7 @@
         <spring-security.version>5.7.3</spring-security.version>
         <spring-boot-framework.version>2.7.4</spring-boot-framework.version>
         <spring-boot-ws-framework.version>1.4.7.RELEASE</spring-boot-ws-framework.version>
+        <spring-cloud.version>3.2.7</spring-cloud.version>
         <spring-native.version>0.12.1</spring-native.version>
 
         <!--Springdoc-->
@@ -364,11 +365,16 @@
                 <version>1.19.0</version>
             </dependency>
 
-            <!-- Spring GCP-->
+            <!-- Spring Cloud Adapters-->
+            <dependency>
+                <groupId>org.springframework.cloud</groupId>
+                <artifactId>spring-cloud-function-adapter-aws</artifactId>
+                <version>${spring-cloud.version}</version>
+            </dependency>
             <dependency>
                 <groupId>org.springframework.cloud</groupId>
                 <artifactId>spring-cloud-function-adapter-gcp</artifactId>
-                <version>3.2.7</version>
+                <version>${spring-cloud.version}</version>
             </dependency>
 
             <!--Spring Dependencies-->
@@ -528,6 +534,11 @@
             </dependency>
 
             <!--Spring Cloud-->
+            <dependency>
+                <groupId>org.springframework.cloud</groupId>
+                <artifactId>spring-cloud-starter-function-web</artifactId>
+                <version>${spring-cloud.version}</version>
+            </dependency>
             <dependency>
                 <groupId>org.springframework.cloud</groupId>
                 <artifactId>spring-cloud-starter-sleuth</artifactId>
@@ -901,6 +912,20 @@
                 <groupId>io.projectreactor</groupId>
                 <artifactId>reactor-core</artifactId>
                 <version>3.4.24</version>
+            </dependency>
+
+            <!--AWS-->
+            <dependency>
+                <groupId>com.amazonaws</groupId>
+                <artifactId>aws-lambda-java-events</artifactId>
+                <version>2.0.2</version>
+                <scope>provided</scope>
+            </dependency>
+            <dependency>
+                <groupId>com.amazonaws</groupId>
+                <artifactId>aws-lambda-java-core</artifactId>
+                <version>1.1.0</version>
+                <scope>provided</scope>
             </dependency>
         </dependencies>
     </dependencyManagement>
